+++ Rakefile
@@ -13,4 +13,4 @@ Dir.glob(File.expand_path('../lib/tasks/**/*.rake', __FILE__)) do |file|
   load file
 end
 
-task :default => [:ci]
+task :default => ['preservation:ci']
