+++ lib/preservation/event_logger.rb
