--- lib/tasks/services.rake
@@ -1,8 +1,0 @@
-require 'preservation/service_environment'
-
-namespace :services do
-  task :start, [:env] do |t, args|
-    env = args[:env] || 'development'
-    Preservation::ServiceEnvironment.new(env).start
-  end
-end
